requirements.txt
cffi==1.15.1
click==8.1.3
file==0.3.0
jmespath==1.0.1
pycparser==2.21
PyYAML==6.0
typer==0.6.1
